Display post on user feed
